refactor(maestro): isolate component prop link matching
